+++ CrudRepos.Persistance/Migrations/DatabaseContextModelSnapshot.cs
@@ -24,11 +24,11 @@ namespace CrudRepos.Persistance.Migrations
 
             modelBuilder.Entity("CrudRepos.Domain.Entities.Users.User", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint");
+                        .HasColumnType("int");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
 
                     b.Property<int>("Age")
                         .HasColumnType("int");
